centered menu items in project list and refactored admin menu
